Update styles, and update studio components
