+++ packages/data-context/graphql/schemaTypes/index.ts
@@ -1,9 +1,9 @@
 /* eslint-disable padding-line-between-statements */
 // created by autobarrel, do not modify directly
 
-export * from './enumTypes'
-export * from './inputTypes'
-export * from './interfaceTypes'
-export * from './objectTypes'
-export * from './scalarTypes'
-export * from './unions'
+export * from './enumTypes/'
+export * from './inputTypes/'
+export * from './interfaceTypes/'
+export * from './objectTypes/'
+export * from './scalarTypes/'
+export * from './unions/'
