+++ packages/reporter/cypress/e2e/unit/app_state.cy.ts
@@ -141,25 +141,6 @@ describe('app state', () => {
     })
   })
 
-  context('#togglePreferencesMenu', () => {
-    it('toggles isPreferencesMenuOpen', () => {
-      const instance = new AppState()
-
-      instance.togglePreferencesMenu()
-      expect(instance.isPreferencesMenuOpen).to.be.true
-      instance.togglePreferencesMenu()
-      expect(instance.isPreferencesMenuOpen).to.be.false
-    })
-
-    it('sets reset value for autoScrollingEnabled', () => {
-      const instance = new AppState()
-
-      instance.togglePreferencesMenu()
-      instance.reset()
-      expect(instance.autoScrollingEnabled).to.be.true
-    })
-  })
-
   context('#setStudioActive', () => {
     it('sets studioActive', () => {
       const instance = new AppState()
