+++ packages/reporter/src/lib/app-state.ts
@@ -4,7 +4,6 @@ import { observable, makeObservable } from 'mobx'
 interface DefaultAppState {
   isPaused: boolean
   isRunning: boolean
-  isPreferencesMenuOpen: boolean
   nextCommandName: string | null | undefined
   pinnedSnapshotId: number | string | null
   studioActive: boolean
@@ -16,7 +15,6 @@ interface DefaultAppState {
 const defaults: DefaultAppState = {
   isPaused: false,
   isRunning: false,
-  isPreferencesMenuOpen: false,
   nextCommandName: null,
   pinnedSnapshotId: null,
   studioActive: false,
@@ -29,7 +27,6 @@ class AppState {
   isSpecsListOpen = false
   isPaused = defaults.isPaused
   isRunning = defaults.isRunning
-  isPreferencesMenuOpen = defaults.isPreferencesMenuOpen
   nextCommandName = defaults.nextCommandName
   pinnedSnapshotId = defaults.pinnedSnapshotId
   studioActive = defaults.studioActive
@@ -45,7 +42,6 @@ class AppState {
       isSpecsListOpen: observable,
       isPaused: observable,
       isRunning: observable,
-      isPreferencesMenuOpen: observable,
       nextCommandName: observable,
       pinnedSnapshotId: observable,
       studioActive: observable,
@@ -97,10 +93,6 @@ class AppState {
 
   toggleSpecList () {
     this.isSpecsListOpen = !this.isSpecsListOpen
-  }
-
-  togglePreferencesMenu () {
-    this.isPreferencesMenuOpen = !this.isPreferencesMenuOpen
   }
 
   setSpecsList (status: boolean) {
