+++ packages/reporter/src/studio/StudioTest.tsx
@@ -1,4 +1,4 @@
-import React, { useMemo, useRef } from 'react'
+import React, { useCallback, useMemo, useRef } from 'react'
 import { observer } from 'mobx-react'
 import { RunnablesStore } from '../runnables/runnables-store'
 import { Duration } from '../duration/duration'
@@ -8,9 +8,11 @@ import Tooltip from '@cypress/react-tooltip'
 import cx from 'classnames'
 import Attempts from '../attempts/attempts'
 import { useScrollIntoView } from '../lib/useScrollIntoView'
-import { IconChevronDownSmall, IconStatusFailedSolid, IconStatusPassedSolid, IconStatusQueuedOutline, IconStatusRunningOutline } from '@cypress-design/react-icon'
+import { IconArrowLeft, IconChevronDownSmall, IconStatusFailedSolid, IconStatusPassedSolid, IconStatusQueuedOutline, IconStatusRunningOutline } from '@cypress-design/react-icon'
 import Test from '../test/test-model'
 import { StatsStore } from '../header/stats-store'
+import Button from '@cypress-design/react-button'
+import events from '../lib/events'
 
 const getConnectors = (num: number) => {
   let connectors: JSX.Element[] = []
@@ -70,6 +72,12 @@ export const StudioTest = observer(({ appState, runnablesStore, statsStore }: St
     isStudioActive: appState.studioActive,
   })
 
+  const handleBackButton = useCallback((e: React.MouseEvent<HTMLElement>) => {
+    e.preventDefault()
+
+    events.emit('studio:cancel', undefined)
+  }, [])
+
   // Call callbackAfterUpdate when mounted and test changes
   React.useEffect(() => {
     if (isMounted && currentTest) {
@@ -86,21 +94,26 @@ export const StudioTest = observer(({ appState, runnablesStore, statsStore }: St
       <div className='studio-single-test-container' >
         <div className='studio-header__test-section'>
           <div className='studio-header__test-section-left'>
-            <StatusIcon test={currentTest} />
-            {parentTitles.length > 0 ? (
-              <Tooltip title={<ul className='studio-tooltip__breadcrumb-list' ref={tooltipRef}>
-                {getParentTitlesListElements(parentTitles)}
-              </ul>}
-              wrapperClassName='studio-header__test-tooltip-wrapper' className={cx(
-                'studio-tooltip cy-tooltip',
-              )}>
-                {testTitle}
-              </Tooltip>
-            ) : testTitle}
+            <Button data-cy='studio-back-button' size='32' variant='outline-dark' className='studio-header__back-button' onClick={handleBackButton}>
+              <IconArrowLeft size='16' strokeColor='gray-500' />
+            </Button>
+            <div className='studio-header__test-section-left-content'>
+              <StatusIcon test={currentTest} />
+              {parentTitles.length > 0 ? (
+                <Tooltip title={<ul className='studio-tooltip__breadcrumb-list' ref={tooltipRef}>
+                  {getParentTitlesListElements(parentTitles)}
+                </ul>}
+                wrapperClassName='studio-header__test-tooltip-wrapper' className={cx(
+                  'studio-tooltip cy-tooltip',
+                )}>
+                  {testTitle}
+                </Tooltip>
+              ) : testTitle}
+            </div>
           </div>
           <div className='studio-header__test-section-right'>
             <Duration duration={statsStore.duration} />
-            <Controls appState={appState} displayPreferencesButton={false} />
+            <Controls appState={appState} />
           </div>
         </div>
         <div className='studio-single-test-attempts' ref={containerRef}>
