--- packages/reporter/src/studio/StudioTestHeader.scss
@@ -1,25 +1,0 @@
-.studio-header {
-  display: flex;
-  flex-direction: column;
-  flex-shrink: 0;
-
-  .studio-header__file-section {
-    display: flex;
-    width: 100%;
-    align-items: center;
-    gap: 13px;
-    padding: 0 16px;
-    border-top: 1px solid $gray-900;
-    border-bottom: 1px solid $gray-900;
-    font-size: 14px;
-    line-height: 20px;
-    color: $gray-500;
-    min-height: $header-height;
-
-    .studio-header__back-button {
-      padding: 0;
-      width: 32px;
-      justify-content: center;
-    }
-  }
-}
